Backend/requirements.txt
Django==4.2.*
djangorestframework>=3.14
pymysql>=1.1.0
cryptography>=41.0.0
django-cors-headers>=4.3
python-dotenv>=1.0
PyJWT>=2.8
djangorestframework-simplejwt>=5.3
